[EX-4923] Reorder mocked responses
