+++ packages/x-components/tests/e2e/cucumber/common-steps.spec.ts
@@ -1,24 +1,6 @@
-import {
-  NextQueriesResponse,
-  PageableRequest,
-  RelatedTagsResponse,
-  SearchByIdResponse,
-  SearchResponse,
-  SuggestionsResponse
-} from '@empathyco/x-adapter';
+import { PageableRequest } from '@empathyco/x-adapter';
 import { Given, Then, When } from 'cypress-cucumber-preprocessor/steps';
 import '../cucumber/global-definitions';
-import {
-  createHierarchicalFacetStub,
-  createNumberRangeFacetStub,
-  createResultStub,
-  createSimpleFacetStub,
-  getNextQueriesStub,
-  getPopularSearchesStub,
-  getQuerySuggestionsStub,
-  getRelatedTagsStub,
-  getResultsStub
-} from '../../../src/__stubs__';
 
 let resultsList: string[] = [];
 
@@ -186,248 +168,6 @@ Then(
     });
   }
 );
-
-Given('a next queries API', () => {
-  cy.intercept('https://api.empathy.co/getNextQueries', req => {
-    req.reply(<NextQueriesResponse>{
-      nextQueries: getNextQueriesStub()
-    });
-  });
-});
-
-Given('a suggestions API', () => {
-  cy.intercept('https://api.empathy.co/getSuggestions', req => {
-    req.reply(<SuggestionsResponse>{
-      suggestions: req.body.query ? getQuerySuggestionsStub('rum') : getPopularSearchesStub()
-    });
-  });
-});
-
-Given('a related tags API', () => {
-  cy.intercept('https://api.empathy.co/getRelatedTags', req => {
-    req.reply(<RelatedTagsResponse>{
-      relatedTags: getRelatedTagsStub()
-    });
-  });
-});
-
-Given('a results API', () => {
-  cy.intercept('https://api.empathy.co/search', req => {
-    req.reply(<SearchResponse>{
-      redirections: [],
-      banners: [],
-      promoteds: [],
-      results: getResultsStub(),
-      spellcheck: '',
-      totalResults: 1,
-      facets: [],
-      partialResults: [],
-      queryTagging: {
-        url: 'https://tagging.empathy.co',
-        params: {}
-      }
-    });
-  }).as('interceptedRawResults');
-});
-
-Given('an ID results API', () => {
-  cy.intercept('https://api.empathy.co/searchById', req => {
-    req.reply(<SearchByIdResponse>{
-      results: getResultsStub()
-    });
-  });
-});
-
-Given('a results API with a known response', () => {
-  cy.intercept('https://api.empathy.co/search', req => {
-    req.reply(<SearchResponse>{
-      banners: [],
-      promoteds: [],
-      redirections: [],
-      partialResults: [],
-      queryTagging: {
-        url: 'https://analytics.com',
-        params: {}
-      },
-      spellcheck: '',
-      facets: [
-        createSimpleFacetStub('brand_facet', createSimpleFilter => [
-          createSimpleFilter('Juguetes deportivos', false, 3),
-          createSimpleFilter('Puzzles', false, 0),
-          createSimpleFilter('Construcción', false, 7),
-          createSimpleFilter('Construye', false, 6),
-          createSimpleFilter('Disfraces', false, 0)
-        ]),
-        createNumberRangeFacetStub('price_facet', createNumberRangeFilter => [
-          createNumberRangeFilter({ min: 0, max: 10 }, false),
-          createNumberRangeFilter({ min: 10, max: 20 }, false),
-          createNumberRangeFilter({ min: 20, max: 30 }, false),
-          createNumberRangeFilter({ min: 30, max: 40 }, false),
-          createNumberRangeFilter({ min: 40, max: 60 }, false),
-          createNumberRangeFilter({ min: 60, max: 100 }, false)
-        ]),
-        createNumberRangeFacetStub('age_facet', createNumberRangeFilter => [
-          createNumberRangeFilter({ min: 0, max: 1 }, false),
-          createNumberRangeFilter({ min: 1, max: 3 }, false),
-          createNumberRangeFilter({ min: 3, max: 6 }, false),
-          createNumberRangeFilter({ min: 6, max: 9 }, false),
-          createNumberRangeFilter({ min: 9, max: 12 }, false),
-          createNumberRangeFilter({ min: 12, max: 99 }, false)
-        ]),
-        createHierarchicalFacetStub('hierarchical_category', createFilter => [
-          ...createFilter('Vehículos y pistas', false, createFilter => [
-            ...createFilter('Radiocontrol', false)
-          ]),
-          ...createFilter('Juguetes electrónicos', false, createFilter => [
-            ...createFilter('Imagen y audio', false)
-          ]),
-          ...createFilter('Educativos', false, createFilter => [
-            ...createFilter('Juguetes educativos', false)
-          ]),
-          ...createFilter('Creativos', false, createFilter => [...createFilter('Crea', false)]),
-          ...createFilter('Muñecas', false, createFilter => [
-            ...createFilter('Peluches', false),
-            ...createFilter('Ropa y accesorios', false),
-            ...createFilter('Playsets', false),
-            ...createFilter('Bebés', false),
-            ...createFilter('Carros', false)
-          ]),
-          ...createFilter('Construcción', false, createFilter => [
-            ...createFilter('Construye', false)
-          ])
-        ])
-      ],
-      results: [
-        createResultStub('LEGO Super Mario Pack Inicial: Aventuras con Mario - 71360', {
-          images: ['https://picsum.photos/seed/1/100/100'],
-          price: {
-            hasDiscount: false,
-            originalValue: 59.99,
-            value: 59.99
-          }
-        }),
-        createResultStub('LEGO Duplo Classic Caja de Ladrillos - 1091', {
-          images: ['https://picsum.photos/seed/2/100/100'],
-          price: {
-            hasDiscount: false,
-            originalValue: 29.99,
-            value: 29.99
-          }
-        }),
-        createResultStub('LEGO City Coche Patrulla de Policía - 60239', {
-          images: ['https://picsum.photos/seed/3/100/100'],
-          price: {
-            hasDiscount: false,
-            originalValue: 11.99,
-            value: 11.99
-          }
-        }),
-        createResultStub('LEGO City Police Caja de Ladrillos - 60270', {
-          images: ['https://picsum.photos/seed/4/100/100'],
-          price: {
-            hasDiscount: false,
-            originalValue: 39.99,
-            value: 39.99
-          }
-        }),
-        createResultStub('LEGO Friends Parque para Cachorros - 41396', {
-          images: ['https://picsum.photos/seed/5/100/100'],
-          price: {
-            hasDiscount: false,
-            originalValue: 11.99,
-            value: 11.99
-          }
-        }),
-        createResultStub('LEGO Creator Ciberdrón - 31111', {
-          images: ['https://picsum.photos/seed/6/100/100'],
-          price: {
-            hasDiscount: false,
-            originalValue: 11.99,
-            value: 11.99
-          }
-        }),
-        createResultStub('LEGO Technic Dragster - 42103', {
-          images: ['https://picsum.photos/seed/7/100/100'],
-          price: {
-            hasDiscount: false,
-            originalValue: 22.99,
-            value: 22.99
-          }
-        })
-      ],
-      totalResults: 7
-    });
-  }).as('interceptedResults');
-});
-
-Given('a results API with {int} results', (resultsLength: number) => {
-  cy.intercept('https://api.empathy.co/search', req => {
-    req.reply(<SearchResponse>{
-      banners: [],
-      promoteds: [],
-      redirections: [],
-      partialResults: [],
-      queryTagging: {
-        url: 'https://analytics.com',
-        params: {}
-      },
-      spellcheck: '',
-      facets: [
-        createSimpleFacetStub('brand_facet', createSimpleFilter => [
-          createSimpleFilter('Juguetes deportivos', false, 3),
-          createSimpleFilter('Puzzles', false, 0),
-          createSimpleFilter('Construcción', false, 7),
-          createSimpleFilter('Construye', false, 6),
-          createSimpleFilter('Disfraces', false, 0)
-        ]),
-        createNumberRangeFacetStub('price_facet', createNumberRangeFilter => [
-          createNumberRangeFilter({ min: 0, max: 10 }, false),
-          createNumberRangeFilter({ min: 10, max: 20 }, false),
-          createNumberRangeFilter({ min: 20, max: 30 }, false),
-          createNumberRangeFilter({ min: 30, max: 40 }, false),
-          createNumberRangeFilter({ min: 40, max: 60 }, false),
-          createNumberRangeFilter({ min: 60, max: 100 }, false)
-        ]),
-        createNumberRangeFacetStub('age_facet', createNumberRangeFilter => [
-          createNumberRangeFilter({ min: 0, max: 1 }, false),
-          createNumberRangeFilter({ min: 1, max: 3 }, false),
-          createNumberRangeFilter({ min: 3, max: 6 }, false),
-          createNumberRangeFilter({ min: 6, max: 9 }, false),
-          createNumberRangeFilter({ min: 9, max: 12 }, false),
-          createNumberRangeFilter({ min: 12, max: 99 }, false)
-        ]),
-        createHierarchicalFacetStub('hierarchical_category', createFilter => [
-          ...createFilter('Vehículos y pistas', false, createFilter => [
-            ...createFilter('Radiocontrol', false)
-          ]),
-          ...createFilter('Juguetes electrónicos', false, createFilter => [
-            ...createFilter('Imagen y audio', false)
-          ]),
-          ...createFilter('Educativos', false, createFilter => [
-            ...createFilter('Juguetes educativos', false)
-          ]),
-          ...createFilter('Creativos', false, createFilter => [...createFilter('Crea', false)]),
-          ...createFilter('Muñecas', false, createFilter => [
-            ...createFilter('Peluches', false),
-            ...createFilter('Ropa y accesorios', false),
-            ...createFilter('Playsets', false),
-            ...createFilter('Bebés', false),
-            ...createFilter('Carros', false)
-          ]),
-          ...createFilter('Construcción', false, createFilter => [
-            ...createFilter('Construye', false)
-          ])
-        ])
-      ],
-      results: Array.from({ length: resultsLength }, (_, index) =>
-        createResultStub(`Result ${index}`, {
-          images: [`https://picsum.photos/seed/${index}/100/100`]
-        })
-      ),
-      totalResults: 24
-    });
-  }).as('interceptedResults');
-});
 
 When('tab is reloaded', () => {
   cy.reload();
