+++ packages/x-components/tests/e2e/cucumber/partial-results/partial-results.spec.ts
@@ -1,62 +1,4 @@
-import { SearchResponse } from '@empathyco/x-adapter';
-import { And, Given, Then, When } from 'cypress-cucumber-preprocessor/steps';
-import { getFacetsStub } from '../../../../src/__stubs__/facets-stubs.factory';
-import { createResultStub } from '../../../../src/__stubs__/results-stubs.factory';
-
-// Background
-Given('a results API with partial results', () => {
-  cy.intercept('https://api.empathy.co/search', req => {
-    req.reply(<SearchResponse>{
-      banners: [],
-      promoteds: [],
-      spellcheck: '',
-      totalResults: 1,
-      queryTagging: {
-        url: 'https://tagging.empathy.co/',
-        params: {}
-      },
-      redirections: [],
-      results: [
-        createResultStub('LEGO Super Mario Pack Inicial: Aventuras con Mario - 71360', {
-          images: ['https://picsum.photos/seed/1/100/100']
-        })
-      ],
-      facets: getFacetsStub(),
-      partialResults: [
-        {
-          query: 'verde azul',
-          results: [
-            createResultStub('Twister', {
-              images: ['https://picsum.photos/seed/30/100/100']
-            }),
-            createResultStub('Juego de Anillas Acuáticas Peces', {
-              images: ['https://picsum.photos/seed/31/100/100']
-            }),
-            createResultStub('Jurassic World Dinosaurio de Ataque Varios Modelos', {
-              images: ['https://picsum.photos/seed/32/100/100']
-            })
-          ],
-          totalResults: 9
-        },
-        {
-          query: 'lego verde',
-          results: [
-            createResultStub('LEGO Classic Ladrillos Creativos Verdes - 11007', {
-              images: ['https://picsum.photos/seed/33/100/100']
-            }),
-            createResultStub('LEGO Creator Grandes Dinosaurios -31058', {
-              images: ['https://picsum.photos/seed/34/100/100']
-            }),
-            createResultStub('LEGO My City Casa Familiar - 60291', {
-              images: ['https://picsum.photos/seed/35/100/100']
-            })
-          ],
-          totalResults: 6
-        }
-      ]
-    });
-  }).as('interceptedPartialResults');
-});
+import { And, Then, When } from 'cypress-cucumber-preprocessor/steps';
 
 // Scenario 1
 Then('at least {int} related results are displayed', (minResultsWithoutPartials: number) => {
