+++ packages/x-components/tests/e2e/cucumber/popular-searches/popular-searches.spec.ts
@@ -1,23 +1,7 @@
-import { SuggestionsResponse } from '@empathyco/x-adapter';
 import { And, Given, Then, When } from 'cypress-cucumber-preprocessor/steps';
-import { createPopularSearch } from '../../../../src/__stubs__';
 import { InstallXOptions } from '../../../../src/x-installer/x-installer/types';
 
 // Background
-And('a popular searches API with a known response', () => {
-  cy.intercept('https://api.empathy.co/getSuggestions', req => {
-    req.reply(<SuggestionsResponse>{
-      suggestions: [
-        createPopularSearch('playmobil'),
-        createPopularSearch('lego'),
-        createPopularSearch('mochila'),
-        createPopularSearch('barbie'),
-        createPopularSearch('dinosaurio')
-      ]
-    });
-  }).as('interceptedPopularSearches');
-});
-
 Given(
   'following config: hide session queries {boolean}, requested items {int}, rendered {int}',
   (hideSessionQueries: boolean, maxItemsToRequest: number, maxItemsToRender: number) => {
