+++ packages/x-components/tests/e2e/cucumber/related-tags/related-tags.spec.ts
@@ -1,34 +1,7 @@
-import { RelatedTagsResponse } from '@empathyco/x-adapter';
 import { And, Given, Then, When } from 'cypress-cucumber-preprocessor/steps';
-import { createRelatedTagStub } from '../../../../src/__stubs__';
 
 let relatedTagsList: string[] = [];
 const secondRelatedTagsList: string[] = [];
-
-// Background
-Given('a related tags API with a known response', () => {
-  cy.intercept('https://api.empathy.co/getRelatedTags', req => {
-    req.reply(<RelatedTagsResponse>{
-      relatedTags: [
-        createRelatedTagStub('lego', 'marvel'),
-        createRelatedTagStub('lego', 'bombero'),
-        createRelatedTagStub('lego', 'policia')
-      ]
-    });
-  }).as('interceptedRelatedTags');
-});
-
-Given('a related tags API with a selected one', () => {
-  cy.intercept('https://api.empathy.co/getRelatedTags', req => {
-    req.reply(<RelatedTagsResponse>{
-      relatedTags: [
-        createRelatedTagStub('lego', 'bombero', { selected: true }),
-        createRelatedTagStub('lego', 'policia'),
-        createRelatedTagStub('lego', 'barbie')
-      ]
-    });
-  }).as('interceptedRelatedTagsWithSelection');
-});
 
 // Scenario 1
 Given(
